+++ src/geom/vector.rs
@@ -71,6 +71,7 @@ impl Vector {
     }
 
     ///Clamp a vector somewhere between a minimum and a maximum
+    #[must_use]
     pub fn clamp(self, min_bound: Vector, max_bound: Vector) -> Vector {
         Vector::new(
             max_bound.x.min(min_bound.x.max(self.x)),
@@ -79,6 +80,7 @@ impl Vector {
     }
     
     ///Constrain a vector within a Rectangle
+    #[must_use]
     pub fn constrain(self, bounds: Rectangle) -> Vector {
         self.clamp(bounds.top_left(), bounds.top_left() + bounds.size())
     }
@@ -94,26 +96,31 @@ impl Vector {
     }
 
     ///Normalize the vector's length from [0, 1]
+    #[must_use]
     pub fn normalize(self) -> Vector {
         self / self.len()
     }
 
     ///Get only the X component of the Vector, represented as a vector
+    #[must_use]
     pub fn x_comp(self) -> Vector {
         Vector::new(self.x, 0f32)
     }
 
     ///Get only the Y component of the Vector, represented as a vector
+    #[must_use]
     pub fn y_comp(self) -> Vector {
         Vector::new(0f32, self.y)
     }
 
     ///Get the vector equal to Vector(1 / x, 1 / y)
+    #[must_use]
     pub fn recip(self) -> Vector {
         Vector::new(self.x.recip(), self.y.recip())
     }
 
     ///Multiply the components in the matching places
+    #[must_use]
     pub fn times(self, other: Vector) -> Vector {
         Vector::new(self.x * other.x, self.y * other.y)
     }
@@ -124,6 +131,7 @@ impl Vector {
     }
 
     ///Create a vector with the same angle and the given length
+    #[must_use]
     pub fn with_len(self, length: f32) -> Vector {
         self.normalize() * length
     }